pyproject.toml
[project]
name = "vibe-analyze"
version = "0.1.0"
description = "High-recall codebase answering CLI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "tiktoken>=0.7.0",
  "PyYAML>=6.0",
  "google-generativeai>=0.7.2",
]

[project.scripts]
vibe-analyze = "vibe_analyze.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

